LTK.Decide.Acom

Description

This module implements an algorithm to decide whether a given FSA has a semigroup both Commutative and Aperiodic (Acom). This is the class that LTT localizes (Almeida, 1989).

https://doi.org/10.1016/0022-4049(89)90124-2

Since: 1.1

Documentation

isAcom :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes a \(\langle 1,t\rangle\)-LTT stringset.

isAcomM :: (Ord n, Ord e) => SynMon n e -> Bool

True iff the monoid is aperiodic and commutative

comTest :: (Ord n, Ord e) => SynMon n e -> Set (State (S [Maybe n] e)) -> Bool

True iff the specified elements commute.