LTK.Decide.LTT

Description

This module implements an algorithm to decide whether a given FSA is Locally Testable (LTT) based on the semigroup characterization of Beaquier and Pin from their 1989 work "Factors of Words".

Since: 0.2

Documentation

isLTT :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes an LTT stringset.

isLTTM :: (Ord n, Ord e) => SynMon n e -> Bool

True iff the monoid recognizes an LTT stringset.

Since: 1.0

isLTTs :: FiniteSemigroupRep s => s -> Bool

True iff the semigroup recognizes an LTT stringset.

Since: 1.2