LTK.Factors

Description

This module provides a means to define positive and negative factors over the adjacency or precedence relations, as well as unions and intersections thereof.

Constructions

required :: Factor e -> Literal e

The factor is required to appear in every string. Note that a conjunctive constraint of (required (Substring x True True)) restricts the stringset to at most one word.

forbidden :: Factor e -> Literal e

The factor is not allowed to appear in any word.

buildLiteral :: (Enum n, Ord n, Ord e) => Set e -> Literal e -> FSA n e

Build an FSA representing a single constraint.

build :: (Enum n, NFData n, Ord n, NFData e, Ord e) => Set e -> Set (Conjunction e) -> FSA n e

Build an FSA representing the conjunction of a set of constraints provided in conjunctive normal form.

makeConstraint :: Ord e => [[Literal e]] -> Conjunction e

Combine inner lists by Disjunction, and form a Conjunction of the results.

Logical Expressions

data Factor e

A substring or subsequence, from which to build constraints.

Constructors

Substring

Fields

substring :: [Set e]

The sequence of symbol types, e.g. [wxs0, wxs0] for two consecutive unstressed syllables.

headAnchored :: Bool

Anchored to the head of the word?

tailAnchored :: Bool

Anchored to the tail of the word?

Subsequence [Set e]

data Literal e

A constraint.

Constructors

Literal Bool (Factor e)

newtype Disjunction e

Multiple constraints, joined by OR.

Constructors

Disjunction (Set (Literal e))

newtype Conjunction e

Multiple disjunctions, joined by AND.

Constructors

Conjunction (Set (Disjunction e))